dumpling: rename extracted goodix fingerprint HAL
The goodix fingerprint HAL is compiled with module id "gf_fingerprint".
Libhardware loads modules depending on their module id, constructs a path
and fails to load.

OnePlus added a commit to libhardware in order to use a specific path for
loading modules.
https://github.com/OnePlusOSS/android_hardware_libhardware/commit/6cf1401f933d9916bf91300b97f8bdafa48b2db3

we do not want to add that, that's why we need to rename the fingerprint HAL
to "gf_fingerprint.msm8998.so" to be able to use "gf_fingerprint" as module id
and satisfy the HAL.

Change-Id: Ie6ad6b26af87a5cbee9e82c3a5c021197b367a1a
Signed-off-by: Alexander Martinz <alex@amartinz.at>